HW06 annotations version 4.
