+++ OtusJavaHW/hw06-annotations_reflection/src/test/java/ru/otus/Annotations/Annotations.java
@@ -1,0 +1,9 @@
+package ru.otus.Annotations;
+
+public enum Annotations {
+	BEFORE_ALL,
+	AFTER_ALL,
+	BEFORE_EACH,
+	AFTER_EACH,
+	TEST;
+}
